backend/requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
redis
rq
pyserial
python-dotenv
gunicorn